Refactored to pull commonalities of Mono Alphabet Ciphers into an abstract super class.
