+++ package/src/Cryptography/CaesarCipher.java
@@ -1,13 +1,6 @@
 package Cryptography;
 
-class CaesarCipher {
-
-    private static final String plainTextAlphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-    private static final int alphaLength = plainTextAlphabet.length();
-    private String cipherTextAlphabet = plainTextAlphabet.toLowerCase();
-
-    CaesarCipher() {
-    }
+class CaesarCipher extends MonoAlphaCipher {
 
     void setKey(int key) {
         int shift = key % alphaLength;
@@ -18,20 +11,5 @@ class CaesarCipher {
                 plainTextAlphabet.substring(shift, alphaLength) +
                         plainTextAlphabet.substring(0, shift)
         ).toLowerCase();
-    }
-
-    String encode(String plainText) {
-        String cipherText = plainText.toUpperCase();
-        for (int i = 0; i < alphaLength; i++) {
-            cipherText = cipherText.replace(plainTextAlphabet.charAt(i), cipherTextAlphabet.charAt(i));
-        }
-        return cipherText.toUpperCase();
-    }
-
-    String decode(String encodedText) {
-        String decodedText = encodedText.toLowerCase();
-        for (int i = 0; i < alphaLength; i++)
-            decodedText = decodedText.replace(cipherTextAlphabet.charAt(i), plainTextAlphabet.charAt(i));
-        return decodedText;
     }
 }
