+++ package/src/Cryptography/KeywordSwapCipher.java
@@ -1,10 +1,6 @@
 package Cryptography;
 
-class KeywordSwapCipher {
-
-    private static final String plainTextAlphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-    private static final int alphaLength = plainTextAlphabet.length();
-    private String cipherTextAlphabet = plainTextAlphabet.toLowerCase();
+class KeywordSwapCipher extends MonoAlphaCipher {
 
     void setKeyword(String keyword) {
         String tempCipher = "";
@@ -24,20 +20,5 @@ class KeywordSwapCipher {
 
         cipherTextAlphabet = tempCipher.toLowerCase();
 
-    }
-
-    String encode(String plainText) {
-        String cipherText = plainText.toUpperCase();
-        for (int i = 0; i < alphaLength; i++) {
-            cipherText = cipherText.replace(plainTextAlphabet.charAt(i), cipherTextAlphabet.charAt(i));
-        }
-        return cipherText.toUpperCase();
-    }
-
-    String decode(String encodedText) {
-        String decodedText = encodedText.toLowerCase();
-        for (int i = 0; i < alphaLength; i++)
-            decodedText = decodedText.replace(cipherTextAlphabet.charAt(i), plainTextAlphabet.charAt(i));
-        return decodedText;
     }
 }
